Support login without any authentication
This **MUST NOT** be used in production, this is used for testing or
demoing the app.

When activated, this allows a user to login by simply visiting the
app. A fake token is created and saved into the Play session. The
session will be read by the server on subsequent requests.

For this to work, the following some settings need to be set in
`application.conf` or via environment variables. See
`GarageNoLoginSpec.scala` to check which settings need to be changed.

Secure cookies need to be disabled, or the app needs to be running
with https. If using unsecure cookies, you'll need to change
`WS_PROTOCOL` and possibly other settings to make this work.

The dummy access token is signed using the secret stored in
`application.conf` or via the `JWT_SECRET` environment variable. This
means that the token can be used to authenticate against other
services that are using local verification without auth+. This is
typically set by using `AUTH_VERIFICATION=local`.