Use custom version of play-zipkin-tracing
This is a bit unfortunate.

There is a bug in `play-zipkin-tracing` that causes a
NullPointerException.

A bug report was submitted but not yet merged. Therefore a version of
this library with the fix applied was published to
`nexus.advancedtelematic.com/`. As soon as that fixed in merged and
official version is published we can bump that version here.

Signed-off-by: Simão Mata <simao.mata@here.com>