Clean up `TokenVerification` logic
All the logic for token verification was either trivial (returning always
`true`) or not used (class `TokenIntrospection`). So we can simply remove
it. We only need to check the validity of the access token inside the
`IdentityAction` to not make a useless request to Here OIDC when we know
the response will be 401.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>