Remove unused NamespaceFromIdentity.
This classed was used when we used to have namespace = userId. But that
is not the case since f1be3c56 (see OTA-2618).
I can't imagine we'll ever want to support this case, now that we have
the whole organizations/environments logic. Moreover, we're not using
this in any environment: I've checked the old hats, EKS and DEV/SIT/PRD.
So there's no use for this, and we can clean it up.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>