Add a route to check if the signing keys are online.
Rationale

When a user uploads an image through the UI, if the keys
are offline, the image uploads normally but it fails at
the end. It fails because after the upload the BE tries
to sign the targets but it doesn't have the signing key.
If the file is big, the user has to wait a lot only to
have it fail at the end.

We want to show an early warning to the user, saying that
the upload will fail and maybe even disallowing it completely
if the keys are offline. For that we need a route to find
out if the keys are offline.

Relates-to: OTA-4764
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>