Pass commit message to deployment-descriptors.
If there is a Jira ticket in the commit message, e.g. OTA-123,
the deployment-descriptors deployments will parse the ticket
and add a comment with the link to the pipeline.

Note that for that to work, the ticket needs to be in the merge-
commit message. Gitlab sometimes does that automatically if the
ticket is mentioned in the commits that are part of the MR, but
sometimes it won't. You might have to add the ticket manually to
the merge-commit message, or you can also consider using the
"Include merge request description" checkbox in the Gitlab MR view.

Relates-to: OTA-4821
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>