Bypass Sonar coverage errors.
So out of nowhere, today building the Sonar coverage reports
are failing with:

```
[error] Not authorized. Analyzing this project requires to be authenticated. Please provide the values of the properties sonar.login and sonar.password.
3045 [error] (sonarScan) Not authorized. Analyzing this project requires to be authenticated. Please provide the values of the properties sonar.login and sonar.password.
3046 [error] Total time: 1 s, completed Aug 4, 2020 7:31:06 AM
```

That hinders us from deploying the app, because the test
stage fails. But that used to work until today. I don't know
what's the cause of that, but until we find out, we can split
the tests and coverage stages (as it's done in all other
services, incidentally). Since the coverage stage is allowed
to fail, at least that will unblock the pipelines and permit
us to deploy the app.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>