Add a `NamespaceProvider` that will only allow existing users.
This alternative implementation of the `NamespaceProvider` trait will
only allow users to the portal if they already have an OTA account in
our system. It will reject users that have a Here account but have never
yet logged in to OTA Connect.

Also, hide the button to register for a new Here Account in the login
page. Make it configurable so that we can still show it and hide it
easily later on.

Relates-to: OTA-5177
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>